+++ packages/primeng/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/ngx-form-engine-primeng.component';
 export * from './lib/ngx-json-form-stepper.component';
 export * from './lib/ngx-json-form-debug.component';
 export * from './lib/signature-pad.component';
+export * from './lib/time-slots.component';